requirements.txt
python-telegram-bot==20.8
schedule==1.2.0
asyncio==3.4.3
